# bytes_pools with buf ptr