@micln 2020-01-04
benchmark 

before

```
Running tool: /usr/local/bin/go test -benchmem -run=^$ github.com/dubbogo/gost/bytes -bench ^(BenchmarkAcquireBytesSize32|BenchmarkAcquireBytesSize10k|BenchmarkAcquireBytesSize60k|BenchmarkAcquireBytesSize70k|BenchmarkFindIndexSize8|BenchmarkFindIndexSize60k)$ -vet=off -v

goos: darwin
goarch: amd64
pkg: github.com/dubbogo/gost/bytes
BenchmarkAcquireBytesSize32-8    	15031885	        72.4 ns/op	      32 B/op	       1 allocs/op
BenchmarkAcquireBytesSize10k-8   	15527878	        76.5 ns/op	      32 B/op	       1 allocs/op
BenchmarkAcquireBytesSize60k-8   	15352584	        78.8 ns/op	      32 B/op	       1 allocs/op
BenchmarkAcquireBytesSize70k-8   	  186849	      7063 ns/op	   73728 B/op	       1 allocs/op
BenchmarkFindIndexSize8-8        	408673654	         3.06 ns/op	       0 B/op	       0 allocs/op
BenchmarkFindIndexSize60k-8      	150715929	         7.66 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/dubbogo/gost/bytes	8.629s
Success: Benchmarks passed.
```

after

```
Running tool: /usr/local/bin/go test -benchmem -run=^$ github.com/dubbogo/gost/bytes -bench ^(BenchmarkAcquireBytesSize32|BenchmarkAcquireBytesSize10k|BenchmarkAcquireBytesSize60k|BenchmarkAcquireBytesSize70k|BenchmarkFindIndexSize8|BenchmarkFindIndexSize60k)$ -vet=off -v

goos: darwin
goarch: amd64
pkg: github.com/dubbogo/gost/bytes
BenchmarkAcquireBytesSize32-8    	14476160	        72.8 ns/op	      32 B/op	       1 allocs/op
BenchmarkAcquireBytesSize10k-8   	15807932	        76.9 ns/op	      32 B/op	       1 allocs/op
BenchmarkAcquireBytesSize60k-8   	15221932	        79.0 ns/op	      32 B/op	       1 allocs/op
BenchmarkAcquireBytesSize70k-8   	  175810	      7717 ns/op	   73760 B/op	       2 allocs/op
BenchmarkFindIndexSize8-8        	268327411	         4.01 ns/op	       0 B/op	       0 allocs/op
BenchmarkFindIndexSize60k-8      	157281532	         8.22 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/dubbogo/gost/bytes	8.747s
Success: Benchmarks passed.
```